backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
boto3
python-dotenv
